fix step analysis action bug
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@21281 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Controller/src/main/java/org/gusdb/wdk/controller/action/analysis/AbstractStepAnalysisIdAction.java
@@ -21,7 +21,7 @@ public abstract class AbstractStepAnalysisIdAction extends WdkAction {
   
   private static final Map<String,ParamDef> PARAMS =
       new MapBuilder<String, ParamDef>().put(StepAnalysisContext.ANALYSIS_ID_KEY,
-          new ParamDef(Required.REQUIRED, DataType.INTEGER)).toMap();
+          new ParamDef(Required.REQUIRED, DataType.LONG)).toMap();
   
   private StepAnalysisFactory _analysisMgr;
